Add allowAutoIOPSPerGBIncrease to translated AWS EBS StorageClasses
By default, AWS EBS CSI driver does not increase IOPS of a volume to the
minimal value supported by AWS - such increase results in increased costs,
which might be surprising. It returns an error instead (increase IOPSPerGB
or volume size).

In-tree volume plugin does increase the IOPS. In order to preserve this
behavior of volumes migrated to CSI, the translated StorageClasses should
include "allowAutoIOPSPerGBIncrease" option that allows the driver to
increase IOPS as the in-tree volume plugin would do.

Kubernetes-commit: 3bad31b531e6d69a5185e9099116e2c020b583ec