Fix translation of Azure availability zones
Some Azure regions do not have any availability zones. Nodes in such zones have
these labels, where "0" is a fault domain, not an availability zone.

  "topology.kubernetes.io/region": "westus"
  "topology.kubernetes.io/zone": "0"

Azure Disk CSI driver uses zone "" in this case:
  "topology.disk.csi.azure.com/zone": ""

Fix the CSI translation to translate the fault domain "0" to "". Any value that is just
a number and not "<region>-<number>" ("centralus-1") needs to be translated to
"" in similar fashion.

Kubernetes-commit: ac9c93224ee2d4d3370f1d2575f77f304524205b