rbd: try to tolerate the images without 'kubernetes-dynamic-pvc' prefix
The image hash which are dynamically provisioned by the intree driver
would contain this prefix. however if its statically provisioned image
it may not be the case. In such cases while retrieving the hash the
translation library should not try to index the splitted string based
on the assumed prefix.

Also added unit test for image name without dynamic provisioner prefix

Signed-off-by: Humble Chirammal <hchiramm@redhat.com>

Kubernetes-commit: c4b67ed6465209e835c80fed3d7d12218279642e